# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TheDotFactoryChanges - OpenAPI 3.0
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: https://localhost:{port}/api/{basePath}
    variables:
      port:
        enum:
          - '59572'
        default: '59572'
      basePath:
        enum:
          - 'v1'
        default: v1
paths:
...
  /fontNames:
    get:
      tags:
        - font
      summary: Get available font names
      description: Get available font names
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FontName'
        '401':
          description: User is not authorized
        '500':
          description: Server error
  /fonts:
    get:
      tags:
        - font
      summary: Get font
      description: Get font
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Font'
        '401':
          description: User is not authorized
        '500':
          description: Server error
    post:
      tags:
        - font
      summary: Create font
      description: Create font
      requestBody:
        description: Font
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Font'
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad request
        '401':
          description: User not authorized
        '500':
          description: Server error
  /fonts/{fontId}:
    get:
      tags:
        - font
      summary: Get available fonts
      description: Get available fonts
      parameters:
        - name: fontId
          in: path
          description: Font id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Font'
        '401':
          description: User is not authorized
        '500':
          description: Server error
    delete:
      tags:
        - font
      summary: Delete font by id
      description: Delete font by id
      parameters:
        - name: fontId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: User is not authorized
        '404':
          description: Font not found
        '500':
          description: Server error
    put:
      tags:
        - font
      summary: Update font
      description: Update font
      parameters:
        - name: fontId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Font
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Font'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: User not authorized
        '404':
          description: Font not found
        '500':
          description: Server error
...
  /convertionsNames:
    get:
      tags:
        - convertion
      summary: Get convertion names
      description: Get convertion names
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConvertionName'
        '401':
          description: User is not authorized
        '500':
          description: Server error
  /convertions:
    get:
      tags:
        - convertion
      summary: Get convertions
      description: Get convertions
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Convertion'
        '401':
          description: User is not authorized
        '500':
          description: Server error
    post:
      tags:
        - convertion
      summary: Save convertion
      description: Save convertion
      requestBody:
        description: Convertion
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveConvertionRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveConvertionResponse'
        '400':
          description: Bad request
        '401':
          description: User is not authorized
        '500':
          description: Server error
  /convertions/{convertionId}:
    get:
      tags:
        - convertion
      summary: Get convertion by id
      description: Get convertion by id
      parameters:
        - name: convertionId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Convertion'
        '400':
          description: Bad request
        '401':
          description: User is not authorized
        '500':
          description: Server error
  /convertions/from-template:
    get:
      tags:
        - convertion
      summary: Make convertion
      description: Make convertion
      parameters:
        - name: fontId
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: configId
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: template
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoConvertionResponse'
        '400':
          description: Bad request
        '401':
          description: User is not authorized
        '500':
          description: Server error
components:
  schemas:
...
    FontName:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: string
    Font:
      required:
        - id
        - name
        - size
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        size:
          type: integer
          format: int64
...
    ConvertionName:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: string
    Convertion:
      required:
        - id
        - head
        - body
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        head:
          type: string
          example: string
        body:
          type: string
          example: string
    DoConvertionResponse:
      required:
        - head
        - body
      type: object
      properties:
        head:
          type: string
          example: string
        body:
          type: string
          example: string
    SaveConvertionRequest:
      required:
        - name
        - head
        - body
      type: object
      properties:
        name:
          type: string
          example: string
        head:
          type: string
          example: string
        body:
          type: string
          example: string
    SaveConvertionResponse:
      required:
        - id
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10